move weird experimental type commands into separate folder
