+++ src/main/java/frc/robot/commands/experimental/RavisRamseteCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.experimental;
 
 import java.security.DigestInputStream;
 import java.util.ArrayList;
