+++ src/main/java/frc/robot/commands/experimental/waypointCommand.java
@@ -2,42 +2,46 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.experimental;
 
 import java.util.Set;
 
-import edu.wpi.first.math.controller.PIDController;
+//import edu.wpi.first.wpilibj.RobotState;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.robot.subsystems.DiffDriveSubsystem;
 import frc.robot.subsystems.LimelightSubsytem;
+import frc.robot.Robot;
+import frc.robot.RobotState;
 
-public class PointToTagCommand extends CommandBase {
+
+public class waypointCommand extends CommandBase {
   private final LimelightSubsytem limelightSubsytem;
   private final DiffDriveSubsystem diffDriveSubsystem;
   private final Set<Subsystem> subsystems;
-  private final PIDController pid;
-  /** Creates a new PointToTagCommand. */
-  public PointToTagCommand(LimelightSubsytem limelightSubsytem, DiffDriveSubsystem diffDriveSubsystem)
-  {
+  /** Creates a new waypointCommand. */
+  public waypointCommand(LimelightSubsytem limelightSubsytem, DiffDriveSubsystem diffDriveSubsystem) {
+    // Use addRequirements() here to declare subsystem dependencies.
     this.diffDriveSubsystem = diffDriveSubsystem;
     this.limelightSubsytem = limelightSubsytem;
-    this.subsystems = Set.of(this.diffDriveSubsystem, this.limelightSubsytem);
-    pid = new PIDController(0.002, 0.08, 0.001);
-    // Use addRequirements() here to declare subsystem dependencies.
+    this.subsystems = Set.of(this.limelightSubsytem,this.diffDriveSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
     
+      RobotState.setSavedPoint(diffDriveSubsystem.getPose());
+      
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    
-    diffDriveSubsystem.arcadeDrive(0, -pid.calculate(limelightSubsytem.getTx(),0));
+
+
   }
 
   // Called once the command ends or is interrupted.
@@ -47,11 +51,12 @@ public class PointToTagCommand extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return true;
   }
 
   @Override
   public Set<Subsystem> getRequirements() {
       return this.subsystems;
   }
+
 }
